frontend: add new toggle activity composable
This composable is very tiny and meant to be used in combination with
components that are intended/known to be used within a `KeepAlive`
component. In that case the `onActivated` and `onDeactivated` hooks are
available and can be used to act on the "hiding" of the component.

This composable is simply reflecting the current activation state as
a boolean reference variable. Note that it also uses the `onMounted`
hook. This is to allow component to use this composable in both cases,
within `KeepAlive` and outside. In the first case, `onMounted` and
`onActivated` are redundant. But it is necessary for the second case
where the component is basically always active.

The composable gets used for the request dialog and transfer history
tabs to toggle their action buttons.
Furthermore, to avoid any issues with having two buttons displayed
within the action button portal, only the first one will be shown and
all others are hidden.

There are no tests for this composable as it is incredibly hard to do.
I made an attempt and even after a lot of work it was still not working.
The tests looked so complicated that it is questionable if such tests
actually do anything in case of a so simple source.