frontend: add simple action button vue component
So far we have been abusing the FormKit library for cases we definitely
don't need it. Furthermore it has shown that we are in need of
a standardized design of an "action" button. So far this gets used for the
buttons within the action button portal area. It has no fancy things except
a hover and disabled color effect.