frontend: enforce type imports
Since TypeScript 3.8 it is a good practice to use type imports (`import type
...`) where possible. Type imports allow for certain compiler and tooling
optimizations.

This rule is auto-fixable. So it doesn't add extra work to the
developers. Except the ESLint rule adaption, all changes in this commit are
caused by automated fixes.

Note that this got applied to the `relayer` too.