frontend: display error messages in steps
Currently we still use `daisyUI` for the progress steps. Having this, it
wasn't easy to have multiline content in the step label with different
colors. The workaround with an absolute `<div>`, conditional padding
etc. is not optimal. As it is only a visual component, this is okay. In
the long term we should replace this with a custom implementation that
can be tailed 100% to our needs.