frontend: push back usage of external ethers types
This focuses on the `JsonRpcProvider` class. It got removed everywhere except
for the `request-manager` and `fill-manager` modules which actually communicate
with the contracts. So got basically pushed almost to the most outside ring of
our dependencies.
An exception is the `transfer.spec.ts` because here we still need it to create
`JsonRpcSigner` which we can't get rid of easily.

This also continues the transition to make more use of the configured RPC
endpoints instead of using the users wallet connection. So to get the request
fee costs, we no more need the user to connect the wallet first. And in theory
do the costs update interactively when the user switches the source chain
(currently this still changes the wallet network and reload the page).