frontend: recover token allowance usage
Somehow I managed to loose the call of the token allowance. As soon as the last
allowance is no more enough (always for new users), no transfer is possible
anymore.
This simply adds it back, using the still existing utility function, but
integrating it as a new first step of a transfer. First time that our new
architecture really plays it strength, as we can simply add a news step, at
first place, fully backwards compatible and neatly working.