frontend: add recognition of expired requrests
This provides some functionality that allows to track if a request expires.
When a request got made we used to check if there was already a fulfillment
event and if not wait for such for 30 seconds. If there is no event within that
time, the whole transfer "fails" (even it doesn't actually fail).
This changes to a race between two promises. One checks if there was
a fulfillment event and then listens for new events. At the same time we track
continuously if the request (has) expire(d). Which one of both resolves/rejects
first, finishes the execution of the whole transfer. The other open promise gets
canceled and all event listeners are removed.
The logic is written in a manner so that it (mostly) works also for the future
continuation of a reloaded transfer (which we do not support yet).

A quick reasoning about the approach how to recognize if a request expired:
A request expires if the validity period is over and there are no active claims
for this request. The second condition can be easily verified by checking the
data of the request in the contract as of the last block. Any change to the
claims will produce a new block.
The time condition is much harder. The problem is that the clocks of the user
and the roll-up sequencer are not synchronized. There is no way to get the
current time of the sequencer until he produces a new block which includes
a timestamp (we do not cover the topic that sequences can manipulate time to
some extend and time must not be strict linear). For that reason we always need
a block of the sequence to have the guarantee that the validity period is over
for the sequencer too. But sequencers do only produce blocks on need, which
means there can be potential long period of time without a block.
Any approach to make some guessing of the sequencers time, no matter how complex
and good it is, can lead into a problematic state of the application. If the
time estimate is off, and the sequencer is behind the time, this means liquidity
providers can still fulfill the request and place a claim. But for the
application the transfer is over and will not be tracked anymore. It gets shown
as expired and the user tries to withdraw his funds which fails.
As handling such an error state is very complicated and should be avoided by all
means, we only mark a transfer as expired when it is surely known. Also if this
means that depending on the block-time, this can be delayed for quite some time
until we can show the user. So if a request would be valid for 10min, it can
take 10min, 11min or 30min (incredibly unlikely) until the transfer actually
gets marked as expired in the application.