frontend: Add typescript compiler check to linting
My IDE was reporting the tsc errors while the linting (eslint) was not.
It is good to also have tsc in the linting to be notified about errors. I also added it to the lint:fix script because it make sense to be notified about tsc errors which cannot be autofixed during the eslint fixing.