frontend: simplify transfer component
The `Transfer` component is quite big as it uses many other components and has
to assemble their properties as they are all protocol unaware. Should mean they
are dump visual components and should never see a transfer instance.
As this component will continue to grow, this commit was doing some work to tidy
it a bit up and shim some lines without loosing readability.

One result of this goal was the new `format` function on the `TokenAmount` that
replaces the former `fromattedAmount` getter. It can be used without parameters
to get the same result as the getter before. Additionally it allows to control
how many decimal places to show and if the token symbol should be added.