frontend: Use Chain and Token objects directly as selector values
When constructing the selector options, we use chain id and token address as the values for our SelectorOptions. The construction is actually done from the Chain and Token objects. Thereby, we lose some information (token decimals, ...) which later needs to get reconstructed by parsing from the configuration. Our selector is capable of using objects as value, which removes a lot of boilerplate code related to reconstructing the token and chain objects.