frontend: rework asynchronous task helper
The most important change is that it now also includes a reference to
a possible error of the task.
Furthermore is the `active` state correctly handled with `finally`
block.
Moreover this has been converted to be a full composable, as it has all
the characteristics of it and gets used to compose new composables.