frontend: add custom tooltip component
In regards of getting rid of some unnecessary dependencies (in this case
DaisyUI), a new `Tooltip` component has been added. The advantage of this
component is that is allows for tons of custom features and 100% adaption to
designer requests.
This pure visual component is quite complex in terms of placing and styling.
Unfortunately it is with the current unit test environment only partially
possible to test all features.