frontend: Add development config
This adds a config file for connecting to the local ganache test chain. Since our documentation provides a deterministic way to spawn a chain and deploy the contracts, I think it is beneficial to have a shared config for connecting to it. Otherwise the configuration would be completely broken since the newest contracts are not deployed on a testnet yet. This also changes the target chain selector to be able select the same as the source chain on a development environment. This is necessary since the local development environment uses only one chain.