frontend: adapt dependency free progress component
The DaisyUI dependency can be a pita and we want to get rid of it. Therefore we
need to add custom implementations for the occurrences where we use it.
One of these places were the `Progress` and `ProgressStep` components. They used
the classes of `steps*` to get the list with circles and lines between.
Recently, when we added error message to steps, it was not possible to create
dynamic sizing of the steps. Furthermore we had to trick around with absolute
placement, which lead to accidentally cut text etc. pp. Everything but not
optimal. At this point a dependency is no more helpful but painful.
This commit provides a custom implementation of these steps. The advantage: it
is much less code than before and 100% customizable to our needs. So got the
previously mentioned problems already solved. Any future request by the designer
can be easily addressed too.

Note that this commit merges two commits into one. First of all, it could be
argued that the former split was simply overload. Second it is for the reason
that it makes no sense to have all this special styling of the list items in an
individual list item component. This only makes list for a whole list and
therefore binding this together makes sense. Less code, less tests, better
readability, 100% customizability.