frontend: allow token selection without wallet
When the user has not connected with his wallet yet, the list of selectable
tokens is empty even if he has chosen a source chain. The reason was that the
tokens get filtered by the connected chain of the wallet, not the selected
chain. But the user tends to first put in some values before connecting. Seeing
that no token is available is a bad experience.

This is includes a tiny fix to open the block explorer link in a new tab. This
somehow got lost during merge conflict resolution.