frontend: deactivate transfers on deserialization
If a transfer is still active when the applications gets closed, it will stay
active after reloading the preserved data from the storage. But in fact the
transfer is no more active, there is no process running for them.
Furthermore it is not possible to continue (i.e. trigger `execute` again)
a transfer while it is still marked as being active. This mechanism is important
to prevent any double/concurrent execution of a transfer. But after a restart of
the application there is no other process executing the transfer anymore.
Thereby this commit is also a preparation to allow continuing interrupted
transfers.