frontend: continue interrupted transfers
When a transfer was still active when the application closes, it will now be
continued on the next start of the application. A transfer is considered as
interrupted if was not completed nor has it failed.

To do so the transfer history will be filtered by such transfers and each of
them will be executed again. The multi step action logic will automically
continue from where it was stopped.
As this happens automatically in background, the execution will be triggered
without a connected wallet. This is fine as it would be awkward anyhow else.
Continuing the transfer after the wallet connected can lead to issues because it
is the wrong chain, multiple interrupted transfers can have different chains,
the user gets confrontend with transaction requests by his wallet without
knowing what is going on etc. pp.
In reality it is very unlikely that a transfers fails in the very first few
steps which are the only ones doing a transaction. And if the user quits the
application while being asked to sign a transaction, it should be totally fine
to not continue (in fact retry) this transfer. After all the user did not loose
any funds and the only downside is a failed transfer in the history.