frontend: add proper syncing of past events
When a transfer is awaited to be fulfilled, this includes two parts. We check if
there are any past events we have "missed" that match our request and we listen
to new events. The first part has become very important since interrupted
transfers get continued now. Now it can happen that the user was offline for
a long time and many blocks have been produced in the meantime. If a liquidity
provider has fulfilled the request, we will not find a new event in the future,
but only in the past. But we only synchronize events from 500 blocks in the past.

To resolve this issue we need to synchronize more blocks. In fact all blocks
that got created since the request was made. Only then we can be sure we catch
(or not) the fulfillment event even if the user was offline for ages.

To synchronize from the creation of the request on (and not from 0 or creation
of contracts [which is unknown btw.]) we need to preserve this information (i.e.
the number) within the transfer data. This means before we create a request
transaction on the source chain, we get the current block number of the target
chain. We always do this before the transaction to guarantee we never miss any
event.