frontend: fix incorrect expired and withdraw state
When a request on the source chain has zero active claims and the validity time
has expired, this can mean multiple things. Either the request was actually
filled, but the challenge game took a while. Or it has actually expired. This
means that our current algorithm to determine if a transfer is expired is off.

To resolve this, we need to also check the deposit receiver of the request. If
all properties of expiration are true and nobody has withdrawn it, it expired
and we can withdraw. But if someone else but us has received the deposit
already, it was actually successful. Though, in the latter case we should have
seen a fulfillment event on the target chain first, the promise game could be
lost.

Furthermore we have now additional checks for the withdraw state and function,
fetching data from the blockchain. For example we can now catch the case where
a transfer expired, the user has withdrawn his deposit already and then comes
back to the application. Though we do not subscribe to any events to
continuously update the withdraw state. The user will still see the withdraw
button, but without costs, the user will immediately see that they have been
already withdrawn and that information is remembered.

This includes a good bunch of refactoring in the `transfer` tests. The suite has
become quite big. But it doesn't seem feasible/possible to split the transfer
class yet. But it is possible to reduce the average size of each test case by
outsourcing a lot of boilerplate. This makes the tests better readable and
shorter. Also the general file size could be reduced by that.