frontend: add easy constructor for token amount
Due to encoding restrictions, the constructor is limited to parse encoded data.
That isn't much convenient for regular usage. There is already a static parse
method to create instances based on user input. The new static method just makes
it easier to create instance when an UInt256 instance is known. Note that the
same approach was taken for the `Step` class.