Fix wrong balance displayed when sending a tx
fixes #598

It turns out that the filter for the events is also being executed for "past" events. According to this here: https://github.com/ethers-io/ethers.js/issues/1504 events are emited for the current block.

So what happens is - I execute a transfer, reload the page and my balance is wrong because I received the current block. But my balance was already updated.

My workaround for now is to only fetch the balance from the contract and not to try to calculate it myself.