Fix UI flashing
When submitting a request the UI input was being set to disabled. Unfortunately this caused the UI to "glitch" since the input field and selector dropdown icon backgroudns were white.

By inheriting the background-color from the parent elements the UI stays the same when disabled and one no longer has the feeling that the app is broken.