beamer: don't expect there will be a request if we encounter a claim
We may have ignored a request (e.g. due to an unsupported token pair) so
we need to be able to handle cases where we see claims for requests we
do not have. Note that we do a similar thing in _handle_claim_withdrawn.

Also, while we're here, drop the events that refer to non-existent
requests and claims.