dapp: add option to disable faucet
This feature is necessary for later production deployments where no faucet is
available anyway. With this option set, it will also hide the faucet button at
all. The option is strict and only enables if explicitly set. Any missing or
unknown configuration will disable the faucet. As environment variables (as used
here) are pure strings, the parsing to a boolean is an unnecessary step and
thereby gets skipped.

As this is a build-time configuration and we still plan to maintain a testing
version, a new `staging` mode got introduced. Thereby we have no `development`,
`staging` and `production`. The faucet is enabled for the first to cases.
While the `development` version uses a local chain, `staging` connects to test
chains. This means that for the moment the runtime configuration file for
`staging` and `production` are equal so far, but will diverge in future.