frontend: add inline documentation for core
While we always aim to write self-explanatory code with coded assumptions and
without much comments and external documentation, not all ideas/concepts can be
expressed in code or it is inefficient. Therefore I added some details into the
inline documentation of the most important core classes of the application.
These are meant to include the business logic, implement the backwards
compatibility etc. pp.
I used these notes also in preparation for the code introduction calls with
a new employee and adjusted it based on raised questions. These docs also
include all the information that we put into the issues where this architecture
got discussed.