frontend: remove daisyui dependency
We have removed almost all usages of DaisyUI from the prototype phase. There was
only the imprint modal component which was still using a class of it. Though,
due to how this dependency works, it is hard to guarantee that it got fully
removed.
This marginally reduces the bundle size too.