ci: Remove workflow for PyPI publishing
Currently there is now way to publish a package under a different name
without renaming the underlying module.

As beamer is already taken on PyPI, we have to use a different name.
However, we currently do not know how to do this, and as this is not
important for the mainnet release we decided to postpone this task to a
later point in time.

Ref: https://github.com/beamer-bridge/beamer/issues/843