frontend: fix flashing request fees
This got achieved by simplifying the `useRequestFee` composable. First of all we
don't need to expose a formatted amount anymore. The `amount` value is of type
`TokenAmount` which provides this feature by itself and can be used by any
composable user. No need to duplicate this.
Second the availability value got removed. Due to typing issues we had to
manually check the existence of all paramters in the request function anyway.
Thereby this is kinda duplicated because if the amount value is undefined, this
means it is not available. The advantage is that this remove the small period of
time where the availability was already set but not the amount itself.