frontend: fix connect wallet button
The problem was that if the user connects a wallet, the button to open the menu
gets removed. In context how the button portal works, this means that the `New
Transfer` button gets shown when being in the activity tab.
The actual fix is not optimal, because the portal is not optimal. But it works
reliable. Note that it is important to only hide the button. The CSS ensures
that only the first ported button is shown. The menu button remains there while
being hidden. Before not as it got completely removed.