Add new selector compoenent
The selector uses a new overlay to render all options in a list as it is in the design.

The selector still uses the old SelectorOption<T>. This is because we need to have a uniform way to display a label and an image for any type T. I also considered to only make the output of the modelValue the type T. But this would lead to some problems: We want to be able to set the selector value from the outside, but to get the correct corresponding SelectorOption<T> for displaying it from the props.options we would need to find the value inside the options. This is not possible for an arbitrary type T or would need some interface to allow for a uniform filter for any T that implements that interface. As this is much more effort than the value it would bring, I decided to keep the SelectorOption as the modelValue. Also, the SelectorOption is still in the form.ts file, because we cannot import it from a .vue file.

Lastly, this adds images for the rollups to display in the selector. When we later add images to config of real tokens (in the config files), the images would also be displayed in the selector.