frontend: fix issue of a too low token allowance
In the past, fees where paid in Ether. This has changed, so the fees are now
paid with the same token that gets requested. But the token allowance for the
request manager was still just checked for the amount to transfer. This means in
special cases this check passes, but the allowance is still too low because it
does not fit the additional fees.
The simple fix is just to ensure that the allowance is high enough to cover
transfer amount plus fees.