frontend: Rework switchChain method
I made it simpler by not returning a non-descriptive null value. Instead the function returns a false in case the chain is not added to the provider. True is for the successful case and an error in any other. This is still not perfect and might be properly addressed by an error handling module later. But this at least removes the weird mixed return type null | boolean , which was only confusing.

See this https://github.com/beamer-bridge/beamer/pull/388#discussion_r829551407 for the initial discussion about the behavior.