beamer: improve debug logging
Remove logging that is unnecessary in the vast number of cases so that
we do not require gigabytes of storage when using the debug log level.
Also promote some log messages to the info level, in particular those
that arise out of the agent's actions (e.g. when the agent fills a request
or makes a claim/challenge; actions by other agents continue to be logged
using the debug log level since those are considered less important).