frontend: Make chain switches safely returning the success and sync source chain with provider
Fixes https://github.com/beamer-bridge/beamer/issues/852

This moves the `switchChain` function from the `useChainSelection` composable to the provider (now `switchChainSafely`) and adds a check to ensure that the chain was switched. This is used for the withdraw of a transfer where an unsuccessful switch may cause a tx sent to a wrong chain. Same could happen in the `RequestDialog`. There it is fixed by using a computed property for the source chain selector which stays in sync with the provider's connected chain. If the selector triggers a new value the `switchChainSafely` is called and only after a successful switch the value will be updated.