frontend: Fix transfer history lazy loading
Every transfer group label and every transfer element is now wrapped in LazyWrapper itself. This fixes the problem that a user needed to scroll to the label to see the whole group. Also, this removes the overhead of every transfer being added to the DOM when the respective group is displayed. As all transfers go eventually into the "older" group, this could have been an performance issue in my opinion.