frontend: Extend input field implementation The implementation of the TextInput component has been adjusted to accept other type attributes rather then having a static `text` attribute. Therefore its name has changed as well since it has to serve more purposes than being a text input only.
This fixes #887 - an issue that makes the app unresponsive/unusable after the user has selected the `token` & `sourceChain` and then types an alphabet character inside the `amount` input field (by accident or intentionally).
It also delivers better UX on mobile devices by showing the numeric keyboard to the user for the number inputs.

The Input component still receives/emits the value as string in order to not break current code where `TokenAmount` parsers are being used for serializing the request payload.
Can be changed in future based on needs.

Tests are added for all the basic test cases for the Input component.