beamer: config: support loading TOML configuration files
We start with a default configuration, which must be completed via
command-line options, or a config file, or a combination of both.
Before we hand over the Config object to the rest of the agent,
we make sure that the configuration is sane, i.e. that all the required
keys are set.