frontend: Implement debounced fee calculation & adjust UI fixes #895
Extended implementation of `useRequestFee` composable to support
debounced fee calculation in order to provide better UX and to tackle
the "choppy" reaction when the fees are being updated on the UI.
By adding this we also reduce the stress on the RPC endpoints significantly.

As a side effect, the Spinner component was refactored and another composable
`useDebouncedTask` was created in order to split the implementation logic
for the debounced functionality.

The Spinner component was refactored to provide better documentation and
configurability. It is now used in the request form to indicate to the user
that a fee recalculation is ongoing whenever he changes the amount of tokens
for sending.

Added tests accordingly for the `useRequestFee` composable, `Spinner` component
& `useDebouncedTask` composable.