frontend: Switch to previously selected chain after connecting
This fixes a bug where a user could not select a source chain when they were unconnected, because the source selector was in sync with the provider chain. Now the user can select a chain before connecting and the dapp triggers a switch to the selection when the user connects. Also adds a validity check if the connected chain is the same as the source chain in order to be safe to not create a transfer for the wrong chain (e.g. the connected chain is not supported).