frontend: Delay request expiry listener
Fixes #741

In order to avoid RPC endpoint spam there was a need of an adaptive
transfer expiry event listening.
Previously, the transfer expiry check was attached and called each
time a new block was produced. That is, from the moment a transfer
was created and until it expires we would have this listeners active
for each transfer stored in the `localStorage` that is still pending.
With this change we will use adaptive listeners attached once
the transfer enters the time-frame where it is actually expected
to expire.
There are a few possible scenarios that are covered with this impl.
1. If transfer didn't enter expected expiration time-frame when
checking against sequencer clock && our clock
- We sleep until the transfer enters the expected expiration
time-frame by our clock
2. If transfer entered expected expiration time-frame when checking
against our clock but not against the sequencer clock
- We wait for the sequencer clock to update with the newest
blocks by waiting on and listening on the next few block mints
3. If transfer entered expected expiration time-frame when checking
against our clock && sequencer clock:
3.1 If request was already withdrawn
- Cancel all listeners and return a rejection to signal a
transfer expiry.
3.2 If request still has active claims
- Listen to `ClaimStakeWithdrawn` events until `activeClaims`
falls down to 0. Once 0, return a rejection in order to signal
a transfer expiry.
4. If transfer entered expected expiration time-frame when checking
against our clock and sequencer clock && no one withdrawn the funds
yet && there are no active claims
- Return a rejection in order to signal a transfer expiry
With this approach, we avoid listening and querying the RPC endpoints
unnecessarily by using a more adaptive approach based on our
prior knowledge about the transfers expected behavior throughout
its life-span.