relayer: use Promise.race instead of Promise.any
The latter will block until one of the promises is fulfilled, or all are
rejected. Since the returnOnPpidChange never rejects, this means the
relayer will be forever stuck in case the main promise rejects (i.e.
throws). Promise.race will instead return as soon as one promise is
rejected, which is what we want here.