frontend: Implement config mixer script
There was a need for automated generation of
the config files which are used to configure the dapp.
The config files contain all the necessary information
related to the contract deployed addresses,
chain metadata, tokens supported on each chain, and tokens metadata.
This update brings the logic for mixing the
contract deployment informations with the local
token/chain metadata.
Important notes about the implementation:
- `Tokens` and `Chains` that are not found inside the `config`
folder will be ignored in the mixing process / output and therefore not
visible on the dapp
- When a `Chain` is found inside the `config/chains` folder but a token
related to that `Chain` is not found, the token will be ignored and not
linked to the `Chain` in the final output