frontend: Fix failing transfers
Seems like the code where the `getTransaction` function was used
was not properly checking if the transaction was actually mined in
a block or indexed yet
therefore making the app crash due to returned `null` values
inside the corresponding functions.
From official ethersjs docs its clear that the calls to `getTransaction`
method are not always returning the values we are expecting currently.
When network is congested or for any other reason, mining the blocks
might take longer and that needs to be properly accounted for
when trying to fetch the transactions from the blockchain.
In this case, i switched to using a call to `waitForTransaction` first
with at least 1 confirmation, which will partially ensure us that the
transaction was mined.
By removing the call to `transaction.wait()` we avoid the possible error
exception for when the value of the `transaction` property is null
which was mostly making the app crash unpredictably.