frontend: Fix inconsistent display of transactions
Fixes #965
Due to the nature of how vue dom element `ref` works, there
was an issue when displaying new transactions inside the
transfer history due to accessing certain dom properties that can be
set to `0` if the program haven't waited long enough for the dom element
to appear on the screen.
The problem was occuring when accessing the property `clientHeight` of
the transaction item inside the transaction history list. Even thought
`onMounted` hook was used to ensure that the dom element is properly
attached to the document body, that was not enough since `clientHeight`
will stay 0 until element is fully rendered on the screen.
A simple timeout of 1 millisecond is solving this issue since that
pushes the execution of the function to the end of the call stack
therefore ensuring that the `clientHeight` will be set when the function
gets executed.
Note: for some reason a call to `nextTick` was not enough to make this
work.