frontend: Add onboarding when MM is not installed
Fixes #937
The MetaMask connect button was enabled and not doing anything when
MetaMask extension is not installed in the browser.
With this update we integrate a MetaMask onboarding user flow
for users that haven't installed the MetaMask browser extension
but still want to access the app with it.
The decision to use an onboarding process instead of just disabling
the button was made in order to improve UX by retaining the user flow.