frontend: Add close on wallet menu backdrop click
The only way to close the wallet menu was by using the "close"
indicator shown on the top right of the wallet menu container.
Since the icon is too small and not so prominent i added
another way of closing the menu.
It follows an intuitive approach where the click on the backdrop
of the menu container triggers a close action.