contracts: rework the restricted call mechanism
We now handle everything inside a single contract, RestrictedCalls.
The IMessenger interface got a new function, callAllowed which is
supposed to say whether the call is allowed or not. This completely
pushes call checking to messengers, which are chain-specific.
For same-chain calls, the messengers are not consulted since we can
check directly whether the caller is allowed to make the call.