frontend: new strategy for checking withdraw state
withdrawInfo struct has been removed from the contracts and a
withdrawClaimId property was introduced.
This property will be set to 0 as long as the request haven't been
withdrawn and set to a number other then 0 once it is withdrawn.
By using this logic we can determine if a request has
expired at the time of querying.
Simpler approach but still not optimal.