frontend: Conditionally position tooltips
This makes the tooltip only showing on the right of the main container if the tooltip would fit inside the viewport. If not it is shown below the content element. Also ensures that the left side of the tooltip is in the viewport by setting the left css rule to a minimum of 0. Before the tooltip component was highly generic. It was possible to show it on any side of the main container (while some positions were actually buggy). As with this change it is not obvious how it should behave if another side would be the default one, I decided to remove the unused code and made it only fitting for our use case. If for any reason we need the positioning on a different side, we can get the calculations back from the git history.