frontend: recognize internal wallet disconnect
In order to provide a way for the internal EthereumProvider
implementation to communicate with the rest of the app the
EthereumProvider was extended with an EventEmitter implementation.
In this case, we are signaling to our app once a disconnect was
made by the user from inside his wallet application so we can properly
update our app state.
I am not so happy with the current solution as i believe there is some
code duplication but refactoring the composable/ethereum-provider is a
huge task that should be tackled separately (hopefully very soon).