beamer: add a middleware to cache eth_getBlockByNumber calls
Part of a solution to the issue described at
https://github.com/beamer-bridge/beamer/issues/1086#issuecomment-1270735862

We returned cached responses for all calls querying a concrete block
number, if they exist. If the block number is descriptive, e.g. "latest"
or "pending", we do not look into the cache, but still store the
response in the cache.