frontend: tests: useFaucet composable
Provided tests covering all the `useFaucet` composable functionalities
as well as a small refactor.
I found the passing of a `signer` as a param to be unnecessary since
the composable only needs the wallet address used as a target address
for directing the faucet funds to.
There should be no need of a complete signer object since we dont use
any of the signing functionalities of that object here.