frontend: refactor useTokenBalance composable
This composable contained more information than what is required
and concerned about things that it shouldn't.
Removed all the redundant reactive properties from it as well as
moved the logic for interacting with the contract.
We should always try to move the logic related to the contract
interactions inside their own services, in this example that is the
`token.ts` service.