frontend: refactor useTokenSelection composable
Same as `useChainSelection` composable, this composable should
only focus on mapping the config token list to a selector option list
instead of unnecessary providing reactive properties which are then
directly used inside the form inputs.
Form input reactive properties, if possible should be kept in
the components (or stores) and passed as parameters to composables
instead of the other way around in order to limit the exposure
to side effects.
This also leads to less confusion, more readable & understandable tests
and composables.
In the end, the result of moving parts of the code from the composable
into the component led to way less code and self-contained logic
in both the composable and component.