frontend: improve "use max token balance" feature
The solution provided is based on a prior knowledge of the contracts
and therefore tightly related to the fee structure & calculation logic
that can be found inside the `RequestManager` contract.
Be aware that, if the logic changes inside the contracts, the logic
has to be adjusted here too.

Ideally, we should eagerly look forward to introducing subgraph queries
on the frontend to eliminate the need of constantly querying the rpc
nodes for such features.
Another option for lowering the rpc call count is to completely rewrite
the request-manager service to a class type with cache support.

Slight refactoring was done in order to achieve better testable code.
Provided tests for all the new features/functions.