frontend: implement `NumericInput` component
We had to ensure stability of the `RequestDialog` form by
separating/limiting the functionality scope of the inputs.
Previously, a free-style text input was used for all our needs.
That proved to be a bad idea after many reported critical bugs which
were resulting in a crashing/unresponsive app.
This commit introduces a new input component implementation which
will take care of enforcing the user to insert the expected value
inside the inputs where a numeric value is mandatory.
Mainly, this implementation will make sure that the  `selectedAmount`
model will always hold the correct numeric value which is used in all of
the `BigNumber` parsing/formatting operations.
IOW it removes confusion as well as boosts the UX flow of the form.